Adjust Hours
